rename file #2 - files were missing
