+++ src/packages/editing/services.ts
@@ -1,3 +1,3 @@
 // SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
-export { EditingServiceImpl } from "./EditingImpl";
+export { EditingServiceImpl } from "./EditingServiceImpl";
